Práctico/Resolución de evaluables/TP 5/deliverEats/requirements.txt
Django==3.0.6
django-crispy-forms==1.8.1
Pillow==7.1.2
